refactor : 도메인의 status 기본값 ALIVE로 지정
