+++ aiku/aiku-common/src/main/java/common/domain/team/Team.java
@@ -21,11 +21,11 @@ public class Team extends BaseTime {
     private Long id;
     private String teamName;
 
-    @Enumerated(value = EnumType.STRING)
-    private Status status;
-
     @OneToMany(mappedBy = "team", cascade = CascadeType.ALL)
     private List<TeamMember> teamMembers = new ArrayList<>();
+
+    @Enumerated(value = EnumType.STRING)
+    private Status status = Status.ALIVE;
 
     private Team(String teamName) {
         this.teamName = teamName;
@@ -35,7 +35,6 @@ public class Team extends BaseTime {
     public static Team create(Member member, String teamName){
         //팀 생성
         Team team = new Team(teamName);
-        team.status = Status.ALIVE;
 
         //생성한 멤버를 팀 멤버로 추가
         team.addTeamMember(member);
