+++ aiku/aiku-common/src/main/java/common/domain/team/TeamMember.java
@@ -26,12 +26,11 @@ public class TeamMember extends BaseTime {
     private Member member;
 
     @Enumerated(value = EnumType.STRING)
-    private Status status;
+    private Status status = Status.ALIVE;
 
     protected TeamMember(Team team, Member member) {
         this.team = team;
         this.member = member;
-        this.status = Status.ALIVE;
     }
 
     //==편의 메서드==
