+++ aiku/aiku-main/src/main/java/aiku_main/application_event/event/ScheduleCloseEvent.java
@@ -1,6 +1,6 @@
 package aiku_main.application_event.event;
 
-import common.domain.Schedule;
+import common.domain.schedule.Schedule;
 import common.domain.value_reference.ScheduleValue;
 import lombok.Getter;
