+++ aiku/aiku-main/src/main/java/aiku_main/application_event/publisher/ScheduleEventPublisher.java
@@ -4,8 +4,8 @@ import aiku_main.application_event.event.ScheduleAutoCloseEvent;
 import aiku_main.application_event.event.ScheduleCloseEvent;
 import aiku_main.application_event.event.ScheduleExitEvent;
 import aiku_main.application_event.event.ScheduleOpenEvent;
-import common.domain.Schedule;
-import common.domain.ScheduleMember;
+import common.domain.schedule.Schedule;
+import common.domain.schedule.ScheduleMember;
 import common.domain.member.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
