+++ aiku/aiku-main/src/main/java/aiku_main/dto/ScheduleDetailResDto.java
@@ -1,6 +1,6 @@
 package aiku_main.dto;
 
-import common.domain.Schedule;
+import common.domain.schedule.Schedule;
 import lombok.Getter;
 
 import java.util.List;
