+++ aiku/aiku-main/src/main/java/aiku_main/repository/ScheduleReadRepositoryImpl.java
@@ -6,17 +6,16 @@ import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import common.domain.ExecStatus;
-import common.domain.team.QTeam;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-import static common.domain.QSchedule.schedule;
-import static common.domain.QScheduleMember.scheduleMember;
 import static common.domain.Status.ALIVE;
 import static common.domain.member.QMember.member;
+import static common.domain.schedule.QSchedule.schedule;
+import static common.domain.schedule.QScheduleMember.scheduleMember;
 import static common.domain.team.QTeam.team;
 
 @RequiredArgsConstructor
