+++ aiku/aiku-main/src/main/java/aiku_main/repository/ScheduleRepository.java
@@ -1,12 +1,11 @@
 package aiku_main.repository;
 
 import common.domain.ExecStatus;
-import common.domain.Schedule;
+import common.domain.schedule.Schedule;
 import common.domain.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface ScheduleRepository extends JpaRepository<Schedule, Long>, ScheduleRepositoryCustom {
