+++ aiku/aiku-main/src/main/java/aiku_main/repository/TeamReadRepositoryImpl.java
@@ -6,6 +6,7 @@ import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import common.domain.*;
+import common.domain.schedule.QSchedule;
 import common.domain.team.Team;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
@@ -14,8 +15,8 @@ import java.util.List;
 import java.util.Optional;
 
 import static common.domain.ExecStatus.TERM;
-import static common.domain.QSchedule.schedule;
 import static common.domain.Status.ALIVE;
+import static common.domain.schedule.QSchedule.schedule;
 import static common.domain.team.QTeam.team;
 import static common.domain.team.QTeamMember.teamMember;
