+++ aiku/aiku-main/src/main/java/aiku_main/service/ScheduleService.java
@@ -9,9 +9,9 @@ import aiku_main.repository.ScheduleRepository;
 import aiku_main.repository.TeamRepository;
 import aiku_main.scheduler.ScheduleScheduler;
 import common.domain.ExecStatus;
-import common.domain.ScheduleMember;
+import common.domain.schedule.ScheduleMember;
 import common.domain.member.Member;
-import common.domain.Schedule;
+import common.domain.schedule.Schedule;
 import common.domain.Status;
 import common.domain.team.Team;
 import common.domain.value_reference.TeamValue;
