+++ aiku/aiku-main/src/test/java/aiku_main/integration_test/BettingServiceIntegrationTest.java
@@ -8,6 +8,8 @@ import aiku_main.repository.ScheduleRepository;
 import aiku_main.service.BettingService;
 import common.domain.*;
 import common.domain.member.Member;
+import common.domain.schedule.Schedule;
+import common.domain.schedule.ScheduleMember;
 import common.domain.value_reference.ScheduleMemberValue;
 import common.exception.BaseException;
 import common.exception.NoAuthorityException;
@@ -16,7 +18,6 @@ import common.exception.PaidMemberLimitException;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
