+++ aiku/aiku-main/src/test/java/aiku_main/service/ScheduleServiceTest.java
@@ -10,7 +10,7 @@ import aiku_main.repository.TeamRepository;
 import aiku_main.scheduler.ScheduleScheduler;
 import common.domain.Location;
 import common.domain.member.Member;
-import common.domain.Schedule;
+import common.domain.schedule.Schedule;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
